doing some learning practice

+++ exercises/006-apple_sharing/app.py
@@ -1,6 +1,6 @@
 def apple_sharing(n,k):
   # Your code here
-  return (int(k/n), k%n)
+  return (k/n, k%n)
  
 
 print(apple_sharing(6,50))
